requirements.txt
playwright
snowflake-snowpark-python[pandas]
pytest
black
thefuzz